feat/34-apply-employeecount

+++ src/main/java/com/sprint/project1/hrbank/service/department/DepartmentServiceImpl.java
@@ -6,6 +6,7 @@ import com.sprint.project1.hrbank.dto.department.DepartmentUpdateRequest;
 import com.sprint.project1.hrbank.entity.department.Department;
 import com.sprint.project1.hrbank.mapper.department.DepartmentMapper;
 import com.sprint.project1.hrbank.repository.department.DepartmentRepository;
+import com.sprint.project1.hrbank.repository.employee.EmployeeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,6 +19,7 @@ import java.util.NoSuchElementException;
 public class DepartmentServiceImpl implements DepartmentService {
     private final DepartmentRepository departmentRepository;
     private final DepartmentMapper departmentMapper;
+    private final EmployeeRepository employeeRepository;
 
     @Override
     public DepartmentResponse createDepartment(DepartmentCreateRequest createRequest) {
@@ -26,8 +28,11 @@ public class DepartmentServiceImpl implements DepartmentService {
         department.validateNotDuplicateWith(other);
 
         departmentRepository.save(department);
-        return departmentMapper.toResponse(department, 10);
+        long employeeCount = getEmployeeCountBy(department);
+
+        return departmentMapper.toResponse(department, employeeCount);
     }
+
 
     @Override
     public DepartmentResponse updateDepartment(Long departmentId, DepartmentUpdateRequest updateRequest) {
@@ -36,8 +41,10 @@ public class DepartmentServiceImpl implements DepartmentService {
         Department other = departmentRepository.findByName(updateRequest.name());
         department.validateNotDuplicateWith(other);
 
+        long employeeCount = getEmployeeCountBy(department);
+
         department.update(updateRequest.name(), updateRequest.establishedDate(), updateRequest.description());
-        return departmentMapper.toResponse(department, 10);
+        return departmentMapper.toResponse(department, employeeCount);
     }
 
     @Override
@@ -45,6 +52,15 @@ public class DepartmentServiceImpl implements DepartmentService {
         Department department = departmentRepository.findById(departmentId)
                 .orElseThrow(() -> new NoSuchElementException("department not found" + departmentId));
 
+        long employeeCount = getEmployeeCountBy(department);
+        if (employeeCount > 0) {
+            throw new IllegalStateException("소속 직원이 있는 부서는 삭제할 수 없습니다.");
+        }
         departmentRepository.delete(department);
+    }
+    
+    private long getEmployeeCountBy(Department department) {
+        long employeeCount = employeeRepository.countByDepartmentId(department.getId());
+        return employeeCount;
     }
 }
